+++ framework/domain/src/main/java/is/codion/framework/domain/entity/condition/AbstractCondition.java
@@ -41,7 +41,7 @@ public abstract class AbstractCondition implements Condition, Serializable {
 
 	protected AbstractCondition(EntityType entityType, List<Column<?>> columns, Collection<?> values) {
 		this.entityType = requireNonNull(entityType);
-		this.columns = validate(unmodifiableList(new ArrayList<>(columns)));
+		this.columns = unmodifiableList(new ArrayList<>(columns));
 		this.values = unmodifiableList(new ArrayList<>(values));
 	}
 
@@ -77,18 +77,5 @@ public abstract class AbstractCondition implements Condition, Serializable {
 	@Override
 	public int hashCode() {
 		return Objects.hash(entityType, columns, values);
-	}
-
-	private List<Column<?>> validate(List<Column<?>> columns) {
-		if (!columns.isEmpty()) {
-			for (Column<?> column : columns) {
-				if (!column.entityType().equals(entityType)) {
-					throw new IllegalArgumentException("Condition column entityType mismatch, " +
-									entityType + " expected, got: " + column.entityType());
-				}
-			}
-		}
-
-		return columns;
 	}
 }